Extract OSCALLocations to seperate file
